order_queue/requirements.txt
grpcio>=1.71.0
grpcio-tools>=1.71.0
protobuf>=4.21.0
watchdog==6.0.0